Fix bug that interrupted check cycle and cleanups
The command "for" caused the termination of the check/fix cycle when a
fix on a repository was initiated. By using "while read" this bug has
been fixed.

This commit also includes multiple cleanups of unnecessary code,
including the removal of the universal log which seems to be redundant
after the introduction of the the index page.